requirements.txt
streamlit
yfinance
pandas
plotly
numpy
scipy
streamlit-authenticator
PyYAML
yagmail
